LectorDeCSV y test, no andan pero casi

+++ src/main/java/exceptions/ElPeriodoDeImputacionNoEsValido.java
@@ -1,7 +1,7 @@
 package exceptions;
 
 public class ElPeriodoDeImputacionNoEsValido extends RuntimeException {
-  public ElPeriodoDeImputacionNoEsValido(int lineaConflictiva) {
+  public ElPeriodoDeImputacionNoEsValido(long lineaConflictiva) {
     super("El periodo de imputacion de la linea " + lineaConflictiva + "no es valido");
   }
 }
