+++ src/main/java/exceptions/LaMedicionEsNegativa.java
@@ -1,7 +1,7 @@
 package exceptions;
 
 public class LaMedicionEsNegativa extends RuntimeException {
-  public LaMedicionEsNegativa(int lineaConflictiva) {
+  public LaMedicionEsNegativa(long lineaConflictiva) {
     super("La medicion es negativa en la linea " + lineaConflictiva);
   }
 }
