+++ src/main/java/exceptions/NoSePudoLeerLaLinea.java
@@ -1,7 +1,7 @@
 package exceptions;
 
 public class NoSePudoLeerLaLinea extends RuntimeException {
-  public NoSePudoLeerLaLinea(int lineaConflictiva) {
+  public NoSePudoLeerLaLinea(long lineaConflictiva) {
     super("No se pudo leer la linea nro " + lineaConflictiva + " del archivo");
   }
 }
